Sparse logging and reset
Sparse logging (once every 10 minutes) to not fill memory with logs.
Reset implemented. If the device continuously hits >99.9% RH for
15 minutes, the integral will be set to zero (PID reset). This could
happen e.g. after opening the chamber and venting it so that the
PID-controller would accumulate errors into the integral term while
unable to actually affect the humidity.
This could potentially also cause issues if there are use cases when
this should be allowable.